rrpak: Two-phase env expiration, env locking, more config options.
There was a risk that "rrpak local" and "rrpak run" could choose an old
environment which is almost about to expire. In the worst case it could be
deleted seconds later.

Now temporary environments have both a creation time and a last activation time.
The creation time is only used to decide if "rrpak local" or "rrpak run" will
choose the cached env or download a fresh one. The last activation time decides
if "rrpak clean" is allowed to delete it. The environment is guaranteed to be
available for a configured time from the instant "rrpak local" ran.

Added locks. "rrpak local", "rrpak run", "rrpak cache" should be able to run in
parallel now. "rrpak run" also protects the environment from being deleted while
the command is running, regardless of the configured time.

Added config options for the above. The config setting was renamed because it's
no longer focused on "rrpak update".